requirements.txt
# DocuMind - RAG Document Intelligence System
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sentence-transformers==2.2.2
chromadb==0.4.22
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
pydantic==2.5.3
httpx==0.26.0
numpy==1.26.3
